benefits.dev/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Cors==4.0.0
SQLAlchemy==2.0.23 